gpt_disk_io: Move StdBlockIo into block_io/std_block_io module
